Clean Code Remove unused Class make store &store_type  nullable
